The code was refactored, folders were added, components were placed in the appropriate folders
