+++ src/components/common/table/Table.tsx
@@ -1,9 +1,9 @@
 import React, {ReactNode} from "react";
 import TableHeader from "./TableHeader";
 import TableBody from "./TableBody";
-import {SortByType} from "./UsersList";
-import {ColumnsType} from "./UserTable";
-import {UsersType} from "../api/fake.api/user.api";
+import {SortByType} from "../../page/usersListPage/UsersListPage";
+import {ColumnsType} from "../../ui/UserTable";
+import {UsersType} from "../../../api/fake.api/user.api";
 
 type PropsType = {
     onSort: (item: SortByType) => void
@@ -22,7 +22,7 @@ function Table({onSort, selectedSort, columns, data, children}: PropsType) {
                     <TableBody {...{data, columns}}/>
                 </>
             )}
-        </table>
+        </table> 
     )
 }
