+++ src/components/common/table/index.ts
@@ -1,0 +1,5 @@
+import Table from './Table';
+export default Table;
+import TableBody from './TableBody';
+import TableHeader from './TableHeader';
+export {TableBody, TableHeader};
