+++ src/components/page/userPage/index.ts
@@ -1,0 +1,2 @@
+import UserPage from './UserPage';
+export default UserPage;
