+++ src/components/page/usersListPage/index.ts
@@ -1,0 +1,2 @@
+import UsersListPage from './UsersListPage';
+export default UsersListPage;
