+++ src/components/ui/SearchStatus.tsx
