+++ src/components/ui/qualities/Quality.tsx
