+++ src/components/ui/qualities/index.ts
@@ -1,0 +1,2 @@
+import Qualities from './QualitiesList';
+export default Qualities;
