+++ src/layout/Users.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {useParams} from "react-router-dom";
-import UserPage from "../components/UserPage";
-import UsersList from "../components/UsersList";
+import UserPage from '../components/page/userPage';
+import UsersListPage from '../components/page/usersListPage';
 
 function Users () {
 
@@ -10,7 +10,7 @@ function Users () {
 
     return (
         <>
-            {userId ? <UserPage userId = {userId}/> : <UsersList/>}
+            {userId ? <UserPage userId = {userId}/> : <UsersListPage/>}
         </>
     )
 }
